Transfer the class OffersComponent to new file
