+++ src/components/cards-of-points-of-route.js
@@ -76,7 +76,7 @@ const createMarkUpForPointOfRoute = (point) => {
     `;
 };
 
-export class MarkUpForPointOfRouteComponent extends AbstractComponent {
+export default class MarkUpForPointOfRouteComponent extends AbstractComponent {
   constructor(sameDatePoint) {
     super();
     this._point = sameDatePoint;
@@ -185,36 +185,6 @@ export const renderEventCards = (events, container, onDataChange, onViewChange, 
   }
   return showedControllers;
 };
-
-// create mark-up for offers in list of events:
-const createOffersMarkUp = (offers) => {
-  const offersMarkUp = [];
-  offersMarkUp.push(`<ul class="event__selected-offers">`);
-  for (const offer of offers) {
-    if (offer.isChecked) {
-      offersMarkUp.push(
-          `<li class="event__offer">
-       <span class="event__offer-title">${offer.title}</span>
-        +
-       €&nbsp;<span class="event__offer-price">${offer.price}</span>
-      </li>`
-      );
-    }
-  }
-  offersMarkUp.push(`</ul>`);
-  return offersMarkUp.join(`\n`);
-};
-
-export class OffersComponent extends AbstractComponent {
-  constructor(offers) {
-    super();
-    this._offers = offers;
-  }
-
-  getTemplate() {
-    return createOffersMarkUp(this._offers);
-  }
-}
 
 /* ---------------------------------------------------------------*/
 // time:
